Support FileAbstractSelectionDialog
Answer file names for the fileList.

Pretend to be under the FileDirectory root because otherwise
the dialog will loop endlessly until it finds this root. This is 
because FSPath>>parent answers self if the path is the root
already. And the root paths are not the same as the
FileDirectory root.